Added support for .ldconfig section; Initial partial support of paging added.
.ldconfig or 'link-defined config' is a special section containing config
defined by the results of linking the kernel image. This's useful for the
i386-specific image that has its own linking stage happening before the final
linking stage but needs in runtime information about the kernel image size and
its start location, all of which are defined after the final linking stage.
.ldconfig is basically a structure of data elements which is filled by
the linker via the linker script at the final linking stage.